Now it's working #121
the user only can change, delet or get his own userModel data

+++ EduCrypto/Controllers/UserHandlingController.cs
@@ -1,7 +1,9 @@
 ﻿using Application.Common;
+using Application.Common.Auth;
 using Application.UserHandling;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Configuration;
 using System.Collections.Generic;
 
 namespace EduCrypto.Controllers
@@ -11,14 +13,16 @@ namespace EduCrypto.Controllers
     [Authorize]
     public class UserHandlingController : Controller
     {
+        private IConfiguration config;
         readonly UserHandlingAppService userHandlingAppService;
 
-        public UserHandlingController(ApplicationDbContext dbContext)
+        public UserHandlingController(ApplicationDbContext dbContext, IConfiguration config)
         {
 #if DEBUG
             dbContext.Database.EnsureCreated();
 #endif
             userHandlingAppService = new UserHandlingAppService(dbContext);
+            this.config = config;
         }
 
 #if DEBUG
@@ -29,10 +33,12 @@ namespace EduCrypto.Controllers
         }
 #endif
 
-        [HttpGet("{id}")]
-        [Authorize(Roles = )]
-        public ActionResult GetById(int id)
+        [HttpGet("{userId}")]
+        public ActionResult GetById(int userId)
         {
+            var token = HttpContext.Request.Headers["Authorization"];
+            if (AuthenticationExtension.getUserIdFromToken(config, token) != userId)
+                return Forbid();
             return this.Run(() =>
             {
                 return Ok(userHandlingAppService.GetById(id));
@@ -61,9 +67,11 @@ namespace EduCrypto.Controllers
         [HttpPost]
         public ActionResult Modify(UserHandlingModel user)
         {
+            var token = HttpContext.Request.Headers["Authorization"];
+            if (AuthenticationExtension.getUserIdFromToken(config, token) != user.Id)
+                return Forbid();
             return this.Run(() =>
             {
-                //TODO disable modifing some fields
                 UserHandlingModel modified = userHandlingAppService.GetById(user.Id);
                 modified.Id = user.Id;
                 modified.userName = user.userName;
@@ -88,6 +96,9 @@ namespace EduCrypto.Controllers
         [HttpDelete("{userId}")]
         public ActionResult Delete(int userId)
         {
+            var token = HttpContext.Request.Headers["Authorization"];
+            if (AuthenticationExtension.getUserIdFromToken(config, token) != userId)
+                return Forbid();
             return this.Run(() =>
             {
                 userHandlingAppService.Delete(userId);
